requirements.txt
openai
pandas
streamlit
fastapi
flask
numpy
matplotlib
seaborn
python-dotenv
pyqt6

